extract getiCalEventAsString function to seperate file
